did some reorginzation and fixed lightning
