+++ source/CaveStoryEngine.vcxproj.filters
@@ -279,6 +279,9 @@
     <ClCompile Include="omega.cpp">
       <Filter>Source Files\classes\boss\act</Filter>
     </ClCompile>
+    <ClCompile Include="flash.cpp">
+      <Filter>Source Files\modules\drawing</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="level.h">
@@ -466,6 +469,9 @@
     </ClInclude>
     <ClInclude Include="bossCollision.h">
       <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="flash.h">
+      <Filter>Source Files\modules\drawing</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
